webapp/requirements.txt
-i https://pypi.org/simple
click==6.7
flask_cors==3.0.6
flask-sqlalchemy==2.3.2
flask==1.0.2
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
pyyaml==3.13
sqlalchemy==1.2.11
werkzeug==0.14.1
